Rework Buffers/Mesh
It's still really stupid though, because you can't really deal with
VertexBuffers/IndexBuffers that don't own their data.

I got rid of the virtual function calls and diamond inheritance and now
the buffer knows it's stride (along with the whole vertex format)
and vertex count (previously it was just wrong and very easy to mess up).

The code is simpler and can do a lot more, but you *have to* copy the
data into the VertexBuffer/IndexBuffer. I really want to be able to just
not copy data to them, but rather just set the data of the underlying
Buffer object. But if you do that getCount will be wrong.

It would also be cool if they could update() automatically.