Fix Buffer::subData and add Mesh::updateBuffers
Sadly no automatic updating, but it's already a huge difference to have
updateBuffers. The hard part with auto-updating is figuring out that a
buffer is dirty, which is non-trivial considering that for modification,
I just hand out a reference to the underlying vector<uint8_t> (I don't
want to make it as dirty whenever that getter is called).